- Refactored into a modular structure using blueprints.
